remove regex dependency from file names in comm plot
